Moved device and instance context to separate module.
